linting

+++ provisionersdk/proto/dataupload.go
@@ -99,14 +99,14 @@ func (b *DataBuilder) done() bool {
 
 func BytesToDataUpload(dataType DataUploadType, data []byte) (*DataUpload, []*ChunkPiece) {
 	fullHash := sha256.Sum256(data)
-	size := int64(len(data))
+	size := len(data)
 	// basically ceiling division to get the number of chunks required to
 	// hold the data, each chunk is ChunkSize bytes.
 	chunkCount := int32((size + ChunkSize - 1) / ChunkSize)
 
 	req := &DataUpload{
 		DataHash:   fullHash[:],
-		FileSize:   size,
+		FileSize:   int64(size),
 		Chunks:     chunkCount,
 		UploadType: dataType,
 	}
@@ -115,8 +115,8 @@ func BytesToDataUpload(dataType DataUploadType, data []byte) (*DataUpload, []*Ch
 	for i := int32(0); i < chunkCount; i++ {
 		start := int64(i) * ChunkSize
 		end := start + ChunkSize
-		if end > size {
-			end = size
+		if end > int64(size) {
+			end = int64(size)
 		}
 		chunkData := data[start:end]
